Update all dependency constraints and fix fallout.
Replace the deprecated rollup-plugin-terser with @rollup/plugin-terser.
Then, for all packages in devDependencies, update their dependency
constraint from whatever it was, to ^x.y.0 where x.y is the most
recent published major and minor release of that package, *except*
that for @rollup/plugin-terser, ^0.1.0 is used instead of ^0.2.0
due to a catastrophic bug in v0.2.0 (exactly) for which there is
not yet a fix: https://github.com/rollup/plugins/issues/1366

This fixes a dependency conflict reported by “npm install”:

    npm ERR! While resolving: markdown-it-dollarmath@0.4.2
    npm ERR! Found: eslint-plugin-promise@6.1.1
    npm ERR! node_modules/eslint-plugin-promise
    npm ERR!   dev eslint-plugin-promise@"^6.0.0" from the root project
    npm ERR!
    npm ERR! Could not resolve dependency:
    npm ERR! peer eslint-plugin-promise@"^4.2.1 || ^5.0.0"
    npm ERR!     from eslint-config-standard@16.0.3
    npm ERR! node_modules/eslint-config-standard
    npm ERR!   dev eslint-config-standard@"^16.0.3" from the root project

I believe the update that was actually _required_ to fix this was
eslint-config-standard 16.x -> 17.x, but it is easiest to update
everything at once.

For the peer dependency on markdown-it, change from ^12.3.2 to “12 - 13”
as there do not appear to have been any API changes in that range that
would break this plugin.

These updates required a handful of minor code changes:

 - Rename rollup.config.js to rollup.config.mjs so it is parsed as an
   ES module.
 - Annotate all bundler outputs with ‘exports: "named"’ to squelch
   a new warning (the default export is only for the sake of Jest,
   so the issue described at https://rollupjs.org/guide/en/#outputexports
   should not be relevant to us).
 - The ceremony for importing the terser plugin has changed.
 - The ceremony for getting at katex.renderToString has changed.

With these changes applied, “npm run lint”, “npm run test”, and
“npm run build” all succeed.